Solution to Check if a Number is Prime or Not

+++ Check if a Number is Prime or Not@@6788dd63490637cfc44e1ba2/code.c
@@ -8,7 +8,7 @@ int main(){
         printf("Not Prime");
         return 0;
     }
-    for(int i=2;i<=sqrt(n);i++){
+    for(int i=2;i<=(int)sqrt(n);i++){
         if(n%i==0){
             isPrime=0;
             break;
